Implement inverted template engine
An idea I’ve been thinking about for a while now, a template system
that better suits use cases where the rendering of templates involves a
lot of logic, so the template is actually normal ruby code, while the
rendered content is specified between custom delimiters (here a single
back tick), with support for string interpolation.

On its face it seems a good idea. We’ll soon put it to the test
rendering a wrapper .ly file.